Bloomie (1) copy 4/project_flowers/requirements.txt
Flask==2.2.3
Pillow==10.0.0
numpy==1.24.3
tensorflow-macos==2.16.2
tensorflow-metal==0.8.0
